WorldDreamer/DreamForge/requirements.txt
tensorboard==2.11.2  # pytorch need this. do not use tensorboardX
accelerate==0.20.3
transformers==4.27.4  # for diffusers
hydra-core==1.3.0
einops==0.3.2

Pillow==8.4.0
torchpack==0.3.1
nuscenes-devkit
mpi4py==3.0.3
numpy==1.21.6
numba==0.53.0  # for numpy==1.21.6
llvmlite==0.36.0
opencv-python==4.6.0.66  # visualization requires >=4.5.5.64

moviepy==1.0.3
